% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templateDesigner.R
\name{clean.input}
\alias{clean.input}
\title{Remove whitespace from DNA sequence}
\usage{
clean.input(seq)
}
\arguments{
\item{seq}{A DNA sequence as a string containing ATCG}
}
\value{
TRUE/FALSE stating whether the sequence is valid
}
\description{
This function removes whitespace from the sequence string
}
