% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templateDesigner.R
\name{find.pams}
\alias{find.pams}
\title{Find PAM sites and create sgRNA}
\usage{
find.pams(seq)
}
\arguments{
\item{seq}{A DNA sequence as a string containing ATCG}
}
\value{
a dataframe with the direction of the guide, index and sequence
}
\description{
This scans the DNA sequence for PAM sites and builds sgRNA guides
for the sense and antisense strand
}
