% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templateDesigner.R
\name{get.homology.template}
\alias{get.homology.template}
\title{Create a donor oligo based on input sequence and sgRNA guides}
\usage{
get.homology.template(pam.res, full.seq, ins)
}
\arguments{
\item{pam.res}{The dataframe results after running find.pams()}

\item{full.seq}{A sequence 150bp or longer}

\item{ins}{A DNA sequence that acts as the insert}
}
\value{
a dataframe with the guide sequence and the corresponding donor oligo
}
\description{
This function finds 60bp homology arms up and downstream of the sgRNA cut site and
merges the left homology arm, insert and right homology arm to build a donor template
}
