% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NheI.loxP}
\alias{NheI.loxP}
\title{NheI.loxP}
\format{
A character string:
}
\usage{
NheI.loxP
}
\description{
A short DNA sequence encoding the NheI restriction site and loxp sequence
}
\keyword{datasets}
