% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templateDesigner.R
\name{validate.sequence}
\alias{validate.sequence}
\title{Check sequence}
\usage{
validate.sequence(char.input)
}
\arguments{
\item{char.input}{A DNA sequence as a string containing ATCG}
}
\value{
A boolean value stating whether the sequence is valid
}
\description{
This function inputs a character string and checks that all values
are ATCG
}
